+++ src/presentation/handler/user_handler.py
@@ -2,8 +2,52 @@ from fastapi import Request
 from src.presentation.handler.exchange_auth_app_handler import get_user_info_from_redis_sync
 from src.core.utils import get_redis_adapter
 from src.domain.services.user_service import UserService
+from src.core.logs import debug
 
 user_service = UserService()
+
+
+async def get_fields_info_about_user():
+    """
+    Get fields information about the user table.
+    
+    This endpoint returns the fields (column names and types) of the user table.
+    """
+
+    fields_info = user_service.fields()
+    return fields_info  # Return the fields information
+
+
+async def get_specific_user_info(request: Request, sid: str):
+    """
+    Get specific user information.
+    
+    This endpoint returns user information based on the provided session ID.
+    """
+
+    
+
+    adapter = get_redis_adapter(request)
+
+
+    user_info = await get_user_info_from_redis_sync(adapter=adapter, session_id=sid)
+    
+    # {
+    #     'session_id': 'UFAOWP2XLrJQe4JU-uR1sr1_OgtheU91qQLqhehms_I', 
+    #     'firebase_id': 'fB68zTp1JFaOHWbrTuHav3o03vk2', 
+    #     'access_level': 1, 
+    #     'user_id': 1, 
+    #     'user_uuid_id': 'b7e96850-7be3-4d8e-8bb7-3f2716e29917', 
+    #     'email': 'wilsonmatheuslimaborba@gmail.com', 
+    #     'nonce': 'OJygrObablG-T3-3gV1Y4M91npaqoaaI'
+    # }
+
+    debug(f"User Info: {user_info}")
+
+    user_id = user_info.get("user_id")
+
+    user_data = user_service.get_user_by_id(user_id=user_id)
+    return user_data  # Return the user information
 
 
 async def modify_user_info(request: Request,  sid: str, first_name: str = None, last_name: str = None, phone_number: str = None):